third_party/blink/web_tests/fast/scrolling/scrollbars/scrollbar-mousedown-move-mouseup.html

<!DOCTYPE html>
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<script src='../../../resources/testdriver.js'></script>
<script src='../../../resources/testdriver-actions.js'></script>
<script src='../../../resources/testdriver-vendor.js'></script>
<script src='../../../resources/gesture-util.js'></script>
<script src="../../../resources/blink-coordinates-util.js"></script>
<script src='../../../resources/scrollbar-util.js'></script>

<style>
  #scrollme,
  #notscrollme {
    width: 100px;
    height: 100px;
    overflow: auto;
  }

  #scrollme p {
    height: 1000px;
  }
</style>

<div id='scrollme'>
  <p>This is a scrollable div.</p>
</div>
<div id='notscrollme'></div>

<script>
  function initEventHandlers(element) {
    element.addEventListener('mousedown', handleEvent);
    element.addEventListener('mouseup', handleEvent);
  }

  window.events = {
    'scrollme': [],
    'notscrollme': []
  };

  function handleEvent(e) {
    window.events[e.target.id].push(e);
  }

  window.onload = async () => {
    let d1 = document.querySelector('#scrollme');
    let d2 = document.querySelector('#notscrollme');
    const scroll_start = trackTop(d1);
    const delta_y = d1.getBoundingClientRect().height;

    initEventHandlers(d1);
    initEventHandlers(d2);

    internals.settings.setScrollAnimatorEnabled(false);

    promise_test(async () => {
      if (scrollbarThickness() == 0) {
        // Test is not relevant on platforms that don't display scrollbars.
        return;
      }

      await waitForCompositorReady();
      await pointerDrag(scroll_start.x, scroll_start.y, /*deltaX=*/0, delta_y);

      assert_greater_than(d1.scrollTop, 0);
      assert_equals(d2.scrollTop, 0);
      assert_equals(events['notscrollme'].length, 0);
      assert_equals(events['scrollme'][0].type, 'mousedown');
      assert_equals(events['scrollme'][0].which, 1);
      assert_equals(events['scrollme'][1].type, 'mouseup');
      assert_equals(events['scrollme'][1].which, 1);

      await waitForScrollReset(d1);
      await pointerDrag(scroll_start.x,
        scroll_start.y,
        /*deltaX=*/0,
        delta_y,
        { button: Buttons.MIDDLE });

      assert_greater_than(d1.scrollTop, 0);
      assert_equals(d2.scrollTop, 0);
      assert_equals(events['notscrollme'].length, 0);
      assert_equals(events['scrollme'][2].type, 'mousedown');
      assert_equals(events['scrollme'][2].which, 2);
      assert_equals(events['scrollme'][3].type, 'mouseup');
      assert_equals(events['scrollme'][3].which, 2);
    }, 'Drag a scrollbar and release it on another DIV, only the DIV owns the' +
      ' dragging scrollbar receive mouse events');
  }
</script>
